Don't rely on channel length for queue depth
The length of the channel is a lie. In the dispatch loop, we pull a
job off the queue and block until we can schedule it. This can leave
the job channel empty, returning a queue depth of zero. At that point,
we will accept a new job via Submit() violating our contract.